Move metric report function
